changed phone numb activerecord type to bigint

+++ db/migrate/20231205192829_add_data_to_users.rb
@@ -3,6 +3,6 @@ class AddDataToUsers < ActiveRecord::Migration[7.0]
     add_column :users, :first_name, :string
     add_column :users, :last_name, :string
     add_column :users, :birthday, :date
-    add_column :users, :phone_number, :integer
+    add_column :users, :phone_number, :bigint
   end
 end
